separate internal conv bit functions
